PSARC/2005/740 Tamale Protocol Versioning 6358498 ALOM must generate a response to unknown commands it receives from the host

+++ usr/src/uts/sun4u/sys/rmc_comm_dp.h
@@ -20,7 +20,7 @@
  */
 
 /*
- * Copyright 2002 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -126,7 +126,7 @@ typedef struct dp_req_resp {
 #define	MSG_SENT_BP	0x40
 #define	MSG_RXED_BP	0x80
 
-	uint8_t		error_status;   /* error code */
+	int		error_status;   /* error code */
 
 	uint8_t		retries_left;   /* number of retries left */
